// src/app/precisa/seguridad/perfi-usuarios/model/perfil.ts
export class Perfil{
    constructor(){
    }
    num:            number;
    perfil:         string;
    tipousuario:    number;
    desestado:       string;
    ultimousuario:  string;
    estado:         string;
    UsuarioCreacion?:       string;
    FechaCreacion?:         Date;
    UltimaFechaModif?:         Date;
    ListaPaginas:   any;
}